ai_healthcare_assistant-main/requirements.txt
flask==2.0.1
openai==1.3.0
numpy==1.22.0
scikit-learn==1.0.2
python-dotenv==0.19.2
azure-ai-inference==1.0.0
azure-core==1.26.0